requirements.txt
openai
moviepy
whisper
requests
torch
ffmpeg-python
python-dotenv
